refactor: reorganize markdown document stores
