+++ AngularStandalone/src/app/home/home.component.ts
@@ -3,8 +3,8 @@ import { Store } from '@ngrx/store';
 import { Observable } from 'rxjs';
 
 import { environment } from 'src/environments/environment';
-import { selectDocumentByDocRef } from '../markdown-document/store/markdown-document.selectors';
-import { DocumentRef } from '../store/models/document-ref.model';
+import { selectDocumentByDocRef } from 'src/app/store/document-index/document-index.selectors';
+import { DocumentRef } from 'src/app/store/models/document-ref.model';
 
 @Component({
   selector: 'app-home',
