+++ AngularStandalone/src/app/markdown-document/display/display.component.ts
@@ -1,11 +1,11 @@
 import { ChangeDetectionStrategy, Component, OnDestroy, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
 import { Observable, Subject, takeUntil, delay } from 'rxjs';
-import { BreakpointObserverService } from 'src/app/shared/services/breakpoint-observer.service';
 
+import { BreakpointObserverService } from 'src/app/shared/services/breakpoint-observer.service';
+import { selectDocument } from 'src/app/store/document-index/document-index.selectors';
 import { DocumentRef } from 'src/app/store/models/document-ref.model';
-import { selectFragment } from 'src/app/store/router/router.selector';
-import { selectDocument } from '../store/markdown-document.selectors';
+import { selectFragment } from 'src/app/store/router/router.selectors';
 
 @Component({
   selector: 'app-display',
