+++ AngularStandalone/src/app/shared/utils/lowerCaseComparer.ts
@@ -1,0 +1,3 @@
+export function lowerCaseComparer(a: string, b: string) {
+  return a.toLowerCase().localeCompare(b.toLowerCase());
+}
