+++ AngularStandalone/src/app/store/document-index/document-index-store.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { StoreModule } from '@ngrx/store';
+
+import * as fromDocumentIndex from './document-index.reducer';
+
+@NgModule({
+  imports: [CommonModule, StoreModule.forFeature(fromDocumentIndex.featureKey, fromDocumentIndex.reducer)],
+})
+export class DocumentIndexStoreModule {}
