+++ AngularStandalone/src/app/store/document-index/document-index.actions.ts
@@ -1,0 +1,7 @@
+import { createAction } from '@ngrx/store';
+
+export enum documentIndexActions {
+  loadDocuments = '[Markdown Document Index] load documents',
+}
+
+export const loadDocuments = createAction(documentIndexActions.loadDocuments);
