+++ AngularStandalone/src/app/store/document-search/document-search-store.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { StoreModule } from '@ngrx/store';
+
+import * as fromDocumentSearch from './document-search.reducer';
+
+@NgModule({
+  imports: [CommonModule, StoreModule.forFeature(fromDocumentSearch.featureKey, fromDocumentSearch.reducer)],
+})
+export class DocumentSearchStoreModule {}
