+++ AngularStandalone/src/app/store/models/view-type-options.model.ts
